created a method for updating a row inside of the products table

+++ ORM-Dapper-DB/Products/Product.cs
@@ -7,7 +7,7 @@ namespace ORM_Dapper_DB.Products
         public string Name { get; set; }
         public double Price { get; set; }
         public int CategoryId { get; set; }
-        public byte OnSale { get; set; }
+        public int OnSale { get; set; }
         public string StockLevel { get; set; }
